requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
celery[redis]==5.4.0
redis==5.2.0
pyjwt[crypto]==2.9.0
argon2-cffi==23.1.0
httpx==0.27.2
sse-starlette==2.1.2
pydantic-settings==2.6.1
slowapi==0.1.9
python-multipart==0.0.12

# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0

